* getCenter for PlanarGraph * splitEdge version for default half split added to PlanarGraph * completed the implementation of runSecondaryRoadGeneration
git-svn-id: https://bzflag.svn.sourceforge.net/svnroot/bzflag/trunk@18344 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ bzwgen/inc/FaceGenerator.h
@@ -9,7 +9,7 @@
  * IMPLIED WARRANTIES, INCLUDING, WITHOUT LIMITATION, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
  */
-/** 
+/**
  * @file FaceGenerator.h
  * @author Kornel Kisielewicz kornel.kisielewicz@gmail.com
  * @brief Implements a Generator class that works on graph faces.
@@ -23,22 +23,22 @@
 #include "globals.h"
 #include "Output.h"
 
-/** 
+/**
  * @class FaceGenerator
- * @brief Class for the map generator working on graphs. 
+ * @brief Class for the map generator working on graphs.
  */
 class FaceGenerator : public Generator {
 public:
-  /** 
-   * Constructor, just runs it's inherited constructor. 
+  /**
+   * Constructor, just runs it's inherited constructor.
    */
   FaceGenerator( RuleSet* _ruleset ) : Generator( _ruleset ) {};
-  /** 
-   * Parses options. 
+  /**
+   * Parses options.
    */
   void parseOptions( CCommandLineArgs* /*opt*/ );
   /**
-   * Runs the generator. 
+   * Runs the generator.
    */
   void run( );
   /**
@@ -51,17 +51,17 @@ private:
    * passed value in radians. Assumes that the vector is (0,0) based.
    */
   static Vector2Df deviateVector( const Vector2Df v, double noise );
-  /** 
-   * Main growth function for secondary road generation. Is executed 
-   * recursively. Branching controls how many segments maximum can go 
+  /**
+   * Main growth function for secondary road generation. Is executed
+   * recursively. Branching controls how many segments maximum can go
    * out of the current node (excluding the ones already connected),
    * segmentLength controls the maximum length of a generated segment,
-   * and noise is a catch-all for introducing randomness into the 
-   * generation. 
+   * and noise is a catch-all for introducing randomness into the
+   * generation.
    */
-  void growRoads( graph::Node* node, int branching, double segmentLength, double noise );
-  /** 
-   * Creates the first "face" of the graph, by adding edges and nodes 
+  void growRoads( graph::Node* node, size_t branching, double segmentLength, double noise );
+  /**
+   * Creates the first "face" of the graph, by adding edges and nodes
    * around the world.
    */
   void createInitialGraph( );
@@ -71,13 +71,13 @@ private:
    */
   void runPrimaryRoadGeneration( );
   /**
-   * Creates the layout of the secondary roads. To save on collision 
-   * tests these are stored inside the faces made by the primary 
+   * Creates the layout of the secondary roads. To save on collision
+   * tests these are stored inside the faces made by the primary
    * road network. In case of a normal bzw map, there would be
    * only a few primary roads.
    */
   void runSecondaryRoadGeneration( );
-  /** 
+  /**
    * The final level of road network generation is the subdivision of
    * lots created by secondary road generation into lots acceptable by
    * the ruleset (e.g. with no degenerated faces). This method also is
